Unit testing of thread scheduling API + additional fixes

+++ low_level_platform/impl/src/lf_zephyr_support.c
@@ -157,7 +157,7 @@ int lf_thread_id() { return *((int*)k_thread_custom_data_get()); }
 
 lf_thread_t lf_thread_self() { return k_current_get(); }
 
-int lf_thread_set_cpu(lf_thread_t thread, int cpu_number) { return k_thread_cpu_pin(thread, cpu_number); }
+int lf_thread_set_cpu(lf_thread_t thread, size_t cpu_number) { return k_thread_cpu_pin(thread, cpu_number); }
 
 /**
  * Real-time scheduling API
